merged-app/requirements.txt
flask
flask-socketio
opencv-python>=4.8.0
mediapipe>=0.10.13
numpy>=1.24.0
Pillow>=10.0.0